conditional shouldEndSession + add reprompt
This resolves #22 - about a missing reprompt option for `.play()`. While it is consistent with the docs, it doesen't really create the functionality I was hoping for....

Basically, I added an optional parameter that sets `shouldEndSession` (default True). If you set it to `False` and include a `.reprompt()` clause, then Alexa will enqueue the audio url -- but will pause immediately for user input *before* play audio. Like I said, this is kind of the opposite of what I wanted (prompting for an action at the end of playback)

`audio(text).play(url, offset=0, reprompt=False).simple_card(title, content).reprompt(text)`

If you omit the `reprompt` *argument* in `.play()`, Alexa will play the audio and close the session. This is the default behavior.